// src/app/api/model/groupRead.ts
/**
 * FinanzApp API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { GroupMemberRead } from './groupMemberRead';


/**
 * Representación pública de un grupo en las respuestas de la API.
 */
export interface GroupRead { 
    id: string;
    name: string;
    color: string | null;
    icon: string | null;
    is_active: boolean;
    created_at: string;
    updated_at: string;
    members?: Array<GroupMemberRead>;
}