// src/app/api/model/updateGroupRequest.ts
/**
 * FinanzApp API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Cuerpo de PATCH /account-groups
 */
export interface UpdateGroupRequest { 
    name?: string | null;
    color?: string | null;
    icon?: string | null;
    is_active?: boolean | null;
}